requirements.txt
streamlit
PyMuPDF
google-generativeai
markdown2
xhtml2pdf
fitz
jinja2
PyMuPDF
requests